Keeping code here until I find out how to extract epochs from training history.
